refactor: move tool execution and testing utilities into dedicated modules
