+++ packages/core/ai/src/tools/index.ts
@@ -2,7 +2,7 @@
 // Copyright 2025 DXOS.org
 //
 
-export * from './exec';
+export * from './call';
 export * from './tool';
 export * from './tool-execution-service';
 export * from './tool-resolver-service';
